requirements.txt
websocket
python3-rpi.gpio